+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 5.0.0 (2023-03-28)
+### Breaking Changes
+
+- Field `ID` of struct `VirtualMachineScaleSetIPConfiguration` has been removed
+- Field `ID` of struct `VirtualMachineScaleSetNetworkConfiguration` has been removed
+
+### Features Added
+
+- New field `ProvisionAfterExtensions` in struct `VirtualMachineExtensionProperties`
+- New field `Location` in struct `VirtualMachineScaleSetVMExtension`
+
+
 ## 4.2.0 (2023-03-27)
 ### Features Added